// frontend/TalentTribe/src/app/models/job-seeker-profile.model.ts
export interface JobSeekerProfile {
    jobSeekerProfileId: number;
    fullName?: string;
    skills?: string;
    experience?: string;
    education?: string;
    address?: string;
    city?: string;
    state?: string;
    country?: string;
    postalCode?: string;
    resumeUrl?: string;
    profilePictureUrl?: string;
  }
  